+++ Automate/release-notes.md
@@ -6,6 +6,7 @@
 * Optimized machine scan when a location changes.
 * Fixed bushes in garden pots not automated.
 * Fixed automated trash cans producing different output than they would if checked manually in some cases.
+* Fixed giftbox being automated as a chest.
 
 ## 1.16
 Released 08 March 2020.