+++ Automate/release-notes.md
@@ -7,6 +7,7 @@
 * Fixed bushes in garden pots not automated.
 * Fixed automated trash cans producing different output than they would if checked manually in some cases.
 * Fixed giftbox being automated as a chest.
+* Moved `data.json` into standard `assets` subfolder.
 
 ## 1.16
 Released 08 March 2020.